refactor(GameBoardAnswerForm): move answer form to separate component
